moved functions to the correct file
