+++ openlibrary/core/yearly_reading_goals.py
@@ -1,9 +1,25 @@
-from datetime import datetime
+from datetime import date, datetime
+
+from openlibrary.utils.dateutil import DATE_ONE_MONTH_AGO, DATE_ONE_WEEK_AGO
 from . import db
 
 
 class YearlyReadingGoals:
     TABLENAME = 'yearly_reading_goals'
+
+    @classmethod
+    def summary(cls):
+        return {
+            'total_yearly_reading_goals': {
+                'total': YearlyReadingGoals.total_yearly_reading_goals(),
+                'month': YearlyReadingGoals.total_yearly_reading_goals(
+                    since=DATE_ONE_MONTH_AGO
+                ),
+                'week': YearlyReadingGoals.total_yearly_reading_goals(
+                    since=DATE_ONE_WEEK_AGO
+                ),
+            },
+        }
 
     # Create methods:
     @classmethod
@@ -51,6 +67,22 @@ class YearlyReadingGoals:
             return False
         else:
             return results[0]['current'] >= results[0]['target']
+
+    @classmethod
+    def total_yearly_reading_goals(cls, since: date | None = None) -> int:
+        """Returns a Storage object of <Storage {'count': int}> where 'count' specifies the
+        number reading goals updated. `since` may be used
+        to limit the result to those reading goals updated since a specific
+        date. Any python datetime.date type should work.
+        :param since: returns all reading goals after date
+        """
+        oldb = db.get_db()
+
+        query = f"SELECT count(*) from {cls.TABLENAME}"
+        if since:
+            query += " WHERE updated >= $since"
+        results = list(oldb.query(query, vars={'since': since}))
+        return results[0]
 
     # Update methods:
     @classmethod
